feat: #1342 move validation files to match validation folder convention
